backend/requirements.txt
fastapi[standard]
pydantic
uvicorn
sqlalchemy
pytest
sqlmodel
python-dotenv
